refactor: 微调util
