+++ src/CentralControl.ts
@@ -3,11 +3,19 @@ import * as path from "path";
 import * as express from "express";
 import * as cors from "cors";
 import router from "./route";
-import { logger } from "./Utils";
+import { getAppVersion, logger } from "./util";
 import config from "./Configuration";
 
 
-const appVersion = JSON.parse(fs.readFileSync(path.resolve(`${__dirname}/../package.json`), "utf8")).version;
+/**
+ * 中央控制器
+ */
+export class CentralControl {
+
+}
+
+
+const appVersion = getAppVersion();
 const serverConfig = config.server;
 
 const app = express();
