+++ src/index.ts
@@ -1,2 +1,4 @@
 console.log("Hello, world!");
 console.log("This branch is under reformation with the guide from Class Diagram.");
+
+import "./CentralControl";
